MDEV-32570 Prep: Refactor functions to handle >32-bit lengths
The functions to read row log events from buffers used 32-bit numeric
types to hold the length of the buffer. MDEV-32570 will add in support
for row events larger than what 32-bits can represent, and so this
patch changes the type for the length variable to size_t so larger
row events can be created from raw memory.

+++ sql/log_event.h
@@ -1583,7 +1583,7 @@ public:
     is calculated during write()
   */
   virtual int get_data_size() { return 0;}
-  static Log_event* read_log_event(const uchar *buf, uint event_len,
+  static Log_event* read_log_event(const uchar *buf, size_t event_len,
 				   const char **error,
                                    const Format_description_log_event
                                    *description_event, my_bool crc_check,
@@ -4815,7 +4815,7 @@ protected:
 		 MY_BITMAP const *cols, bool is_transactional,
 		 Log_event_type event_type);
 #endif
-  Rows_log_event(const uchar *row_data, uint event_len,
+  Rows_log_event(const uchar *row_data, size_t event_len,
 		 const Format_description_log_event *description_event);
   void uncompress_buf();
 
@@ -5045,7 +5045,7 @@ public:
                        bool is_transactional);
 #endif
 #ifdef HAVE_REPLICATION
-  Write_rows_log_event(const uchar *buf, uint event_len,
+  Write_rows_log_event(const uchar *buf, size_t event_len,
                        const Format_description_log_event *description_event);
 #endif
 #if defined(MYSQL_SERVER) 
@@ -5095,7 +5095,7 @@ public:
   bool write(Log_event_writer *writer) override;
 #endif
 #ifdef HAVE_REPLICATION
-  Write_rows_compressed_log_event(const uchar *buf, uint event_len,
+  Write_rows_compressed_log_event(const uchar *buf, size_t event_len,
                        const Format_description_log_event *description_event);
 #endif
 private:
@@ -5132,7 +5132,7 @@ public:
   ~Update_rows_log_event() override;
 
 #ifdef HAVE_REPLICATION
-  Update_rows_log_event(const uchar *buf, uint event_len,
+  Update_rows_log_event(const uchar *buf, size_t event_len,
 			const Format_description_log_event *description_event);
 #endif
 
@@ -5184,7 +5184,7 @@ public:
   bool write(Log_event_writer *writer) override;
 #endif
 #ifdef HAVE_REPLICATION
-  Update_rows_compressed_log_event(const uchar *buf, uint event_len,
+  Update_rows_compressed_log_event(const uchar *buf, size_t event_len,
                        const Format_description_log_event *description_event);
 #endif
 private:
@@ -5223,7 +5223,7 @@ public:
   Delete_rows_log_event(THD*, TABLE*, ulonglong, bool is_transactional);
 #endif
 #ifdef HAVE_REPLICATION
-  Delete_rows_log_event(const uchar *buf, uint event_len,
+  Delete_rows_log_event(const uchar *buf, size_t event_len,
 			const Format_description_log_event *description_event);
 #endif
 #ifdef MYSQL_SERVER
@@ -5275,7 +5275,7 @@ public:
   bool write(Log_event_writer *writer) override;
 #endif
 #ifdef HAVE_REPLICATION
-  Delete_rows_compressed_log_event(const uchar *buf, uint event_len,
+  Delete_rows_compressed_log_event(const uchar *buf, size_t event_len,
                        const Format_description_log_event *description_event);
 #endif
 private:
